土地一覧の更新・削除機能を追加

+++ onehouse/database/migrations/2025_10_18_085037_create_onehouse_table.php
@@ -60,7 +60,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('profile_id')->constrained()->cascadeOnDelete();
             $table->string('address');
-            $table->decimal('landarea', 5, 2);
+            $table->integer('landarea');
             $table->unsignedSmallInteger('far');
             $table->unsignedSmallInteger('bcr');
             $table->integer('floor');
